telemetry/otel-demo/app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
requests==2.32.3
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-logging==0.55b1
